removed message_old entity
